Wrote finalpath to .JSON
Made a small start to implementing the python and C# communication. Right now when a finalpath is found it adds the rearwheelpos (x and z), truck heading and hitch angle to a .JSON to be imported in python, this python file succesfully imports the values and stores them in seperate arrays for each value, we can always reconstruct them in the exact way we want after that. This is just to keep them seperate for now which is easier for debugging and handling